+++ XafLayoutBuilder.Sample.Blazor.Server/Startup.cs
@@ -37,6 +37,7 @@ public class Startup {
             builder.Modules
                 .Add<XafLayoutBuilder.Sample.Module.SampleModule>()
                 .Add<XafLayoutBuilder.Blazor.XafLayoutBuilderBlazorModule>()
+                .Add<XafLayoutBuilder.ModelEditor.XafModelEditorModule>()
                 .Add<SampleBlazorModule>();
             builder.ObjectSpaceProviders
                 .AddSecuredEFCore(options => {